combine flows pages into one new one
