+++ demo/sections/flows/index.ts
@@ -2,6 +2,5 @@ import { Section } from '@/demo/router/routeRecords'
 
 export const flows: Section = {
   FlowDetails: () => import('./FlowDetails.vue'),
-  FlowListItem: () => import('./FlowListItem.vue'),
   FlowList: () => import('./FlowList.vue'),
 }
