+++ src/components/FlowList.vue
@@ -1,236 +1,222 @@
 <template>
   <div class="flow-list">
-    <p-layout-table :sticky="media.lg">
-      <template #header>
-        <div class="flow-list__header-container">
-          <div class="flow-list__header">
-            <div class="flow-list__header-start">
-              <template v-if="selected.length == 0">
-                <span v-if="!!flowsCount && !!deploymentsCount" class="flow-list__results-count">
-                  <ResultsCount :label="localization.info.flow" :count="flowsCount" />
-                  {{ localization.info.with }}
-                  <ResultsCount :label="localization.info.deployment" :count="deploymentsCount" />
-                </span>
-              </template>
+    <p-layout-table :root-margin="margin" sticky>
+      <template #header-start>
+        <ResultsCount v-if="selectedFlows.length == 0" label="Flow" :count="total" />
+        <SelectedCount v-else :count="selectedFlows.length" />
 
-              <template v-else-if="selected.length">
-                <SelectedCount :count="selected.length" />
-                <FlowsDeleteButton size="xs" :selected="selected" @delete="deleteFlows" />
-              </template>
-            </div>
-
-            <div class="flow-list__header-end">
-              <SearchInput v-model="search" :placeholder="localization.info.searchByFlowName" :label="localization.info.searchByFlowName" />
-              <SearchInput v-model="deploymentNameSearch" :placeholder="localization.info.searchByDeploymentName" :label="localization.info.searchByDeploymentName" />
-              <p-select v-model="routeFilter.sort" :options="flowSortOptions" />
-              <p-button icon="AdjustmentsVerticalIcon" :class="classes.filterButton" :selected="headerExpanded" @click="headerExpanded = !headerExpanded" />
-            </div>
-          </div>
-
-          <template v-if="headerExpanded">
-            <FlowsFilterGroup />
-          </template>
-          <template v-else-if="isCustomFilter">
-            <div class="flow-list__filters-active">
-              ({{ localization.info.filtersActive }})
-              <p-button small :disabled="isDefaultFilter" @click="clear">
-                {{ localization.info.resetFilters }}
-              </p-button>
-            </div>
-          </template>
-        </div>
+        <FlowsDeleteButton v-if="can.delete.flow" :selected="selectedFlows" small @delete="deleteFlows" />
       </template>
 
-      <template v-if="flowsCount === 0">
-        <p-empty-results>
-          <template #message>
-            {{ localization.info.noFlowsOrDeploymentsMatchFilter }}
-          </template>
-          <template v-if="isCustomFilter" #actions>
-            <p-button small @click="clear">
-              {{ localization.info.resetFilters }}
-            </p-button>
-          </template>
-        </p-empty-results>
+      <template #header-end>
+        <SearchInput v-model="flowNameLike" placeholder="Flow names" label="Search flows" />
+        <FlowRunTagsInput v-model:selected="filter.flowRuns.tags.name" multiple />
       </template>
 
-      <p-virtual-scroller
-        :items="flows ?? []"
-        :chunk-size="20"
-        :item-estimate-height="140"
-        item-key="id"
-      >
-        <template #default="{ item }">
-          <slot v-bind="{ item, selected, selectable, update: handleUpdate, delete: handleDelete }">
-            <FlowListItem
-              v-model:selected="selected"
-              :value="item.id"
-              :flow="item"
-              :filter="routeFilter"
-              :selectable="selectable"
-              class="flow-list__flow"
-              @update="handleUpdate"
-              @delete="handleDelete"
-            />
-          </slot>
+      <template #header-sort>
+        <p-select v-model="filter.sort" :options="flowSortOptions" />
+      </template>
+
+      <p-table :data="flows" :columns="columns">
+        <template #selection-heading>
+          <p-checkbox v-model="model" @update:model-value="selectAllFlows" />
         </template>
-      </p-virtual-scroller>
+
+        <template #selection="{ row }">
+          <p-checkbox v-model="selectedFlows" :value="row.id" />
+        </template>
+
+        <template #name="{ row }">
+          <div class="flow-list__name-col">
+            <p-link :to="routes.flow(row.id)" class="flow-list__name">
+              <span>{{ row.name }}</span>
+            </p-link>
+            <span class="flow-list__created-date">Created {{ formatDateTimeNumeric(row.created) }}</span>
+          </div>
+        </template>
+
+        <template #last-run="{ row }">
+          <LastFlowRun :flow-id="row.id" />
+        </template>
+
+        <template #next-run="{ row }">
+          <NextFlowRun :flow-id="row.id" />
+        </template>
+
+        <template #deployments="{ row }">
+          <DeploymentsCount :flow-id="row.id" />
+        </template>
+
+        <template #activity="{ row }">
+          <MiniFlowHistory
+            class="flow-list__activity-chart"
+            :flow-id="row.id"
+            :time-span-in-seconds="activityTimeSpanInSeconds"
+          />
+        </template>
+
+        <template #action-heading>
+          <span />
+        </template>
+
+        <template #action="{ row }">
+          <div class="flow-list__action">
+            <FlowMenu size="xs" :flow="row" flat @delete="refresh" />
+          </div>
+        </template>
+
+        <template #empty-state>
+          <PEmptyResults>
+            <template #message>
+              No flows
+            </template>
+            <template v-if="isCustomFilter" #actions>
+              <p-button small @click="clear">
+                Clear Filters
+              </p-button>
+            </template>
+          </PEmptyResults>
+        </template>
+      </p-table>
 
       <template #footer-end>
-        <slot name="footer-end" v-bind="{ page, pages }">
-          <p-pager v-if="pages > 1" v-model:page="page" :pages="pages" />
-        </slot>
+        <p-pager v-if="pages > 1" v-model:page="page" :pages="pages" />
       </template>
     </p-layout-table>
   </div>
 </template>
 
 <script lang="ts" setup>
-  import { media } from '@prefecthq/prefect-design'
+  import { PTable, PEmptyResults, PLink, CheckboxModel, TableColumn } from '@prefecthq/prefect-design'
   import { NumberRouteParam, useDebouncedRef, useRouteQueryParam } from '@prefecthq/vue-compositions'
   import merge from 'lodash.merge'
   import { computed, ref } from 'vue'
-  import { FlowListItem, FlowsDeleteButton, ResultsCount, SearchInput, SelectedCount, FlowsFilterGroup } from '@/components'
-  import { useDeploymentsCount, useFlows, useFlowsFilterFromRoute } from '@/compositions'
-  import { localization } from '@/localization'
+  import {
+    FlowsDeleteButton,
+    LastFlowRun,
+    NextFlowRun,
+    DeploymentsCount,
+    ResultsCount,
+    SearchInput,
+    MiniFlowHistory,
+    SelectedCount,
+    FlowRunTagsInput
+  } from '@/components'
+  import { useCan, useFlowsFilterFromRoute, useWorkspaceRoutes, useOffsetStickyRootMargin, useFlows } from '@/compositions'
+  import { useComponent } from '@/compositions/useComponent'
   import { FlowsFilter } from '@/models/Filters'
+  import { Flow } from '@/models/Flow'
   import { flowSortOptions } from '@/types/SortOptionTypes'
+  import { formatDateTimeNumeric } from '@/utilities/dates'
 
   const props = defineProps<{
     filter?: FlowsFilter,
-    selectable?: boolean,
   }>()
 
-  const emit = defineEmits<{
-    (event: 'delete' | 'update', value?: string): void,
-  }>()
+  const activityTimeSpanInSeconds = 24 * 60 * 60
 
-  const headerExpanded = ref(false)
-  const selected = ref<string[]>([])
+  const { margin } = useOffsetStickyRootMargin()
+  const { FlowMenu } = useComponent()
 
-  const search = ref<string>()
-  const searchDebounced = useDebouncedRef(search, 800)
-  const nameLike = computed({
-    get() {
-      return searchDebounced.value === '' ? undefined : searchDebounced.value
-    },
-    set(value: string | undefined) {
-      searchDebounced.value = value
-    },
-  })
+  const can = useCan()
+  const routes = useWorkspaceRoutes()
 
-  const deploymentNameSearch = ref<string>()
-  const deploymentNameLikeDebounced = useDebouncedRef(deploymentNameSearch, 800)
-
-  const deploymentNameLike = computed({
-    get() {
-      return deploymentNameLikeDebounced.value === '' ? undefined : deploymentNameLikeDebounced.value
-    },
-    set(value: string | undefined) {
-      deploymentNameSearch.value = value
-    },
-  })
-
-  const page = useRouteQueryParam<number>('page', NumberRouteParam, 1)
-
-  const { filter: routeFilter, isDefaultFilter, isCustomFilter, clear } = useFlowsFilterFromRoute(merge({}, props.filter, {
+  const flowNameLike = ref<string>()
+  const flowNameLikeDebounced = useDebouncedRef(flowNameLike, 1200)
+  const { filter, clear, isCustomFilter } = useFlowsFilterFromRoute(merge({}, props.filter, {
     flows: {
-      nameLike,
+      nameLike: flowNameLikeDebounced,
     },
-    deployments: {
-      nameLike: deploymentNameLike,
-    },
+    limit: 50,
   }))
 
-  const filter = computed(() => {
-    return {
-      ...routeFilter,
-      limit: 40,
-    }
-  })
-
-  const { subscriptions, flows, total: flowsCount, pages } = useFlows(filter, {
+  const page = useRouteQueryParam('page', NumberRouteParam, 1)
+  const { flows, subscriptions, total, pages } = useFlows(filter, {
     page,
   })
 
-  const { count: deploymentsCount } = useDeploymentsCount(filter)
+  const columns: TableColumn<Flow>[] = [
+    {
+      label: 'selection',
+      width: '20px',
+      visible: can.delete.flow,
+    },
+    {
+      property: 'name',
+      label: 'Name',
+      width: '125px',
+    },
+    {
+      label: 'Last run',
+    },
+    {
+      label: 'Next run',
+    },
+    {
+      label: 'Deployments',
+    },
+    {
+      label: 'Activity',
+    },
+    {
+      label: 'Action',
+    },
+  ]
 
-  const handleUpdate = (flowId: string): void => {
-    emit('update', flowId)
+  const selectedFlows = ref<string[]>([])
+  const selectAllFlows = (allFlowsSelected: CheckboxModel): string[] => {
+    if (allFlowsSelected) {
+      return selectedFlows.value = [...flows.value.map(flow => flow.id)]
+    }
+    return selectedFlows.value = []
+  }
+
+  const model = computed({
+    get() {
+      return selectedFlows.value.length === flows.value.length
+    },
+    set(value: boolean) {
+      selectAllFlows(value)
+    },
+  })
+
+  function refresh(): void {
     subscriptions.refresh()
   }
 
-  const handleDelete = (flowId: string): void => {
-    emit('delete', flowId)
-    subscriptions.refresh()
-  }
+  const emit = defineEmits<{
+    (event: 'delete'): void,
+  }>()
 
   const deleteFlows = (): void => {
+    selectedFlows.value = []
+    refresh()
     emit('delete')
-    subscriptions.refresh()
-    selected.value = []
   }
-
-  const classes = computed(() => ({
-    filterButton: {
-      'flow-list__filter-button--filter-active-text': isCustomFilter.value && !headerExpanded.value,
-      'flow-list__filter-button--filter-open': headerExpanded.value,
-    },
-  }))
 </script>
 
 <style>
-.flow-list {
-  --virtual-scroller-item-gap: theme('spacing.6')
+.flow-list__activity-chart { @apply
+  h-12
+  w-20
 }
 
-.flow-list__header-container { @apply
-  w-full
+.flow-list__action { @apply
+  text-right
 }
 
-.flow-list__header-start { @apply
-  grow
-  whitespace-nowrap
-  items-end
-}
-
-.flow-list__header { @apply
+.flow-list__name-col { @apply
   flex
   flex-col
-  sm:items-center
-  sm:flex-row
-  grow
+  w-80
 }
 
-.flow-list__header-end { @apply
-  flex
-  flex-wrap
-  ml-auto
-  shrink
-  sm:justify-end
-  gap-2
+.flow-list__name { @apply
+  font-medium
 }
 
-.flow-list__results-count { @apply
+.flow-list__created-date { @apply
   text-subdued
-  text-base
-}
-
-.flow-list__filters-active { @apply
-  text-subdued
-  text-sm
-  flex
-  gap-2
-  justify-end
-  items-center
-  mt-4
-}
-
-.flow-list__filter-button--filter-open { @apply
-  cursor-pointer
-}
-
-.flow-list__filter-button--filter-active-text .p-icon { @apply
-  text-selected
+  text-xs
 }
 </style>
