+++ src/components/MiniFlowHistory.vue
@@ -1,0 +1,25 @@
+<template>
+  <FlowRunsBarChart
+    class="flow-list__activity-chart"
+    mini
+    :filter="flowRunsFilter"
+  />
+</template>
+
+<script lang="ts" setup>
+  import { computed } from 'vue'
+  import { FlowRunsBarChart } from '@/components'
+  import { mapper } from '@/services'
+
+  const props = defineProps<{
+    flowId: string,
+    timeSpanInSeconds: number,
+  }>()
+
+  const flowStats = computed(() => ({
+    flowId: props.flowId,
+    timeSpanInSeconds: props.timeSpanInSeconds,
+  }))
+
+  const flowRunsFilter = computed(() => mapper.map('FlowStatsFilter', flowStats.value, 'FlowRunsFilter'))
+</script>
