+++ src/components/index.ts
@@ -75,8 +75,6 @@ export { default as DeploymentDescriptionEmptyState } from './DeploymentDescript
 export { default as DeploymentDetails } from './DeploymentDetails.vue'
 export { default as DeploymentForm } from './DeploymentForm.vue'
 export { default as DeploymentIconText } from './DeploymentIconText.vue'
-export { default as DeploymentList } from './DeploymentList.vue'
-export { default as DeploymentListItem } from './DeploymentListItem.vue'
 export { default as DeploymentMenu } from './DeploymentMenu.vue'
 export { default as DeploymentQuickRunOverflowMenuItem } from './DeploymentQuickRunOverflowMenuItem.vue'
 export { default as DeploymentRelationships } from './DeploymentRelationships.vue'
@@ -96,7 +94,6 @@ export { default as FlowCombobox } from './FlowCombobox.vue'
 export { default as FlowDetails } from './FlowDetails.vue'
 export { default as FlowIconText } from './FlowIconText.vue'
 export { default as FlowList } from './FlowList.vue'
-export { default as FlowListItem } from './FlowListItem.vue'
 export { default as FlowMenu } from './FlowMenu.vue'
 export { default as FlowRouterLink } from './FlowRouterLink.vue'
 export { default as FlowRunArtifacts } from './FlowRunArtifacts.vue'
@@ -115,6 +112,7 @@ export { default as FlowRunList } from './FlowRunList.vue'
 export { default as FlowRunListItem } from './FlowRunListItem.vue'
 export { default as FlowRunLogs } from './FlowRunLogs.vue'
 export { default as FlowRunMenu } from './FlowRunMenu.vue'
+export { default as FlowRunName } from './FlowRunName.vue'
 export { default as FlowRunPauseButton } from './FlowRunPauseButton.vue'
 export { default as FlowRunPauseModal } from './FlowRunPauseModal.vue'
 export { default as FlowRunPopOver } from './FlowRunPopOver.vue'
@@ -141,12 +139,10 @@ export { default as FlowRunTimelineTaskDetails } from './FlowRunTimelineTaskDeta
 export { default as FlowsDeleteButton } from './FlowsDeleteButton.vue'
 export { default as FlowsFilterGroup } from './FlowsFilterGroup.vue'
 export { default as FlowsPageEmptyState } from './FlowsPageEmptyState.vue'
-export { default as FlowsTable } from './FlowsTable.vue'
 export { default as FourZeroFour } from './FourZeroFour.vue'
 export { default as IntervalScheduleForm } from './IntervalScheduleForm.vue'
 export { default as JsonInput } from './JsonInput.vue'
-export { default as ListItemMeta } from './ListItemMeta.vue'
-export { default as ListItemMetaFlowRun } from './ListItemMetaFlowRun.vue'
+export { default as LastFlowRun } from './LastFlowRun.vue'
 export { default as LogLevelLabel } from './LogLevelLabel.vue'
 export { default as LogLevelSelect } from './LogLevelSelect.vue'
 export { default as LogoImage } from './LogoImage.vue'
@@ -154,6 +150,8 @@ export { default as LogRow } from './LogRow.vue'
 export { default as LogsContainer } from './LogsContainer.vue'
 export { default as LogsSort } from './LogsSort.vue'
 export { default as MarketingBanner } from './MarketingBanner.vue'
+export { default as MiniFlowHistory } from './MiniFlowHistory.vue'
+export { default as NextFlowRun } from './NextFlowRun.vue'
 export { default as NotificationDetails } from './NotificationDetails.vue'
 export { default as NotificationForm } from './NotificationForm.vue'
 export { default as NotificationMenu } from './NotificationMenu.vue'
